# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Task Management Portal API
 * OpenAPI spec version: 1.0.0
 */
export interface HealthStatus {
  status: string;
}

export interface MessageResponse {
  message: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface UserOut {
  id: number;
  email: string;
  full_name: string;
  role: string;
  is_active: boolean;
  created_at: string;
}

export interface Token {
  access_token: string;
  token_type: string;
  user: UserOut;
}

export interface UserCreate {
  email: string;
  full_name: string;
  password: string;
  role?: string;
}

export interface UserUpdate {
  email?: string;
  full_name?: string;
  role?: string;
  is_active?: boolean;
  password?: string;
}

export interface StudentOut {
  id: number;
  full_name: string;
  passport_no?: string | null;
  dob?: string | null;
  phone?: string | null;
  email?: string | null;
  created_at: string;
}

export interface StudentCreate {
  full_name: string;
  passport_no?: string;
  dob?: string;
  phone?: string;
  email?: string;
}

export interface StudentUpdate {
  full_name?: string;
  passport_no?: string;
  dob?: string;
  phone?: string;
  email?: string;
}

export interface UniversityOut {
  id: number;
  name: string;
  country?: string | null;
  created_at: string;
}

export interface UniversityCreate {
  name: string;
  country?: string;
}

export interface UniversityUpdate {
  name?: string;
  country?: string;
}

export interface ApplicationOut {
  id: number;
  student_id: number;
  university_id?: number | null;
  assigned_to_id?: number | null;
  created_by_id?: number | null;
  application_status: string;
  assigned_date?: string | null;
  intake?: string | null;
  course?: string | null;
  country?: string | null;
  priority?: string | null;
  source?: string | null;
  remarks?: string | null;
  created_at: string;
  updated_at: string;
  student?: StudentOut | null;
  university?: UniversityOut | null;
  assigned_to?: UserOut | null;
  created_by?: UserOut | null;
}

export interface ApplicationCreate {
  student_id: number;
  university_id?: number | null;
  assigned_to_id?: number | null;
  application_status?: string;
  intake?: string;
  course?: string;
  country?: string;
  priority?: string;
  source?: string;
  remarks?: string;
}

export interface ApplicationUpdate {
  university_id?: number | null;
  assigned_to_id?: number | null;
  application_status?: string;
  intake?: string;
  course?: string;
  country?: string;
  priority?: string;
  source?: string;
  remarks?: string;
}

export interface StatusUpdate {
  application_status: string;
}

export interface AssignUpdate {
  assigned_to_id?: number | null;
}

export interface ActivityLogOut {
  id: number;
  application_id: number;
  field_name: string;
  old_value?: string | null;
  new_value?: string | null;
  changed_at: string;
  changed_by_user?: UserOut | null;
}

export interface DashboardSummary {
  total: number;
  pending: number;
  approved: number;
  refused: number;
}

export interface StatusCount {
  status: string;
  count: number;
  color: string;
}

export interface AssigneeCount {
  assignee_name: string;
  count: number;
}

export interface UniversityCount {
  university_name: string;
  count: number;
}

export interface NotificationTest {
  type: string;
  target: string;
}

export type ListStudentsParams = {
  search?: string;
};

export type ListUniversitiesParams = {
  search?: string;
};

export type ListApplicationsParams = {
  assigned_to_id?: number;
  status?: string;
  search?: string;
};