# lib/api-zod/src/generated/types/applicationCreate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Task Management Portal API
 * OpenAPI spec version: 1.0.0
 */

export interface ApplicationCreate {
  student_id: number;
  university_id?: number | null;
  assigned_to_id?: number | null;
  application_status?: string;
  intake?: string;
  course?: string;
  country?: string;
  priority?: string;
  source?: string;
  remarks?: string;
}