# lib/api-zod/src/generated/types/assigneeCount.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Task Management Portal API
 * OpenAPI spec version: 1.0.0
 */

export interface AssigneeCount {
  assignee_name: string;
  count: number;
}