# lib/api-zod/src/generated/types/statusCount.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Task Management Portal API
 * OpenAPI spec version: 1.0.0
 */

export interface StatusCount {
  status: string;
  count: number;
  color: string;
}