# lib/api-zod/src/generated/types/token.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Task Management Portal API
 * OpenAPI spec version: 1.0.0
 */
import type { UserOut } from "./userOut";

export interface Token {
  access_token: string;
  token_type: string;
  user: UserOut;
}