# lib/api-zod/src/generated/types/universityCount.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Task Management Portal API
 * OpenAPI spec version: 1.0.0
 */

export interface UniversityCount {
  university_name: string;
  count: number;
}