# lib/api-zod/src/generated/types/universityOut.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Task Management Portal API
 * OpenAPI spec version: 1.0.0
 */

export interface UniversityOut {
  id: number;
  name: string;
  country?: string | null;
  created_at: Date;
}