# lib/api-zod/src/generated/types/userOut.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Task Management Portal API
 * OpenAPI spec version: 1.0.0
 */

export interface UserOut {
  id: number;
  email: string;
  full_name: string;
  role: string;
  is_active: boolean;
  created_at: Date;
}